Splitted Main Page into single COmponents
